# api-docs/my-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "植物种植助手小程序后端API文档",
    "description": "植物种植助手小程序后端服务接口文档，基于Spring Boot 2.7.18实现。",
    "contact": {
      "name": "zszleon",
      "email": "2079973876@qq.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:18080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "TokenAuth": []
    }
  ],
...
  "paths": {
...
    "/api/auth/updateProfile": {
      "post": {
        "tags": [
          "认证管理"
        ],
        "summary": "更新用户资料",
        "description": "更新用户昵称和头像等信息，需要Token认证",
        "operationId": "updateProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
...
    "/api/auth/getProfile": {
      "post": {
        "tags": [
          "认证管理"
        ],
        "summary": "获取用户资料",
        "description": "获取当前登录用户的资料信息，需要Token认证",
        "operationId": "getProfile",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResponseObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        },
        "description": "统一API响应"
      },
...
      "ProfileRequest": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "description": "用户昵称",
            "example": "张三"
          },
          "avatarUrl": {
            "type": "string",
            "description": "用户头像URL",
            "example": "https://picx.zhimg.com/v2-abed1a8c04700ba7d72b45195223e0ff_l.jpg?source=1def8aca"
          }
        },
        "description": "更新用户信息请求"
      },
      "ApiResponseProfileResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/ProfileResponse"
          }
        },
        "description": "统一API响应"
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "用户ID",
            "example": "123456789"
          },
          "openid": {
            "type": "string",
            "description": "微信openid",
            "example": "o6_bmjrPTlm6_2sgVt7hMZOPfL2M"
          },
          "nickname": {
            "type": "string",
            "description": "用户昵称",
            "example": "张三"
          },
          "avatarUrl": {
            "type": "string",
            "description": "用户头像URL",
            "example": "https://thirdwx.qlogo.cn/..."
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          }
        },
        "description": "用户信息响应"
      },
...
    },
    "securitySchemes": {
      "TokenAuth": {
        "type": "apiKey",
        "description": "用户认证Token，通过微信登录接口获取",
        "name": "X-Token",
        "in": "header"
      }
    }
  }
}